+++ CHANGELOG.rst
@@ -2,6 +2,26 @@
 Tigger Changelog
 ================
 
+1.6.1
+=====
+
+* Version bump in preparation for release
+* Supports Ubuntu 22.04
+* Beta support for Ubuntu 22.04 ARM64
+* Added FITS header preview pane to file dialog
+* Dependent on the latest tigger-lsm (1.7.1)
+* Fixed float errors with updated library API's
+* Fixed dockable widgets and window sizing
+* Refactored plot zooming
+* Re-enabled splash screen
+* Fixes venv pip3 bug
+* Fixes issue #131
+* Install script 'setup.py install' has been replaced by pip
+* Updated URL's
+* GitHub Actions are now compatible with 'act'
+* Actions now test VENV installations on Ubuntu 20.04 and 22.04
+* Improved Action tests and logging
+
 1.6.0
 =====